Fixed bug #166
Review URL: https://codereview.appspot.com/302290043 .

+++ rekall-core/rekall/utils.py
@@ -483,7 +483,7 @@ class HexDumpedString(AttributedString):
     """A string which should be hex dumped."""
 
 
-class HexInteger(int):
+class HexInteger(long):
     """An int which should be rendered as a hex digit."""
